rpc: move Ensure* helpers to server_util.h
